Introducing the new Flowee Pay solution
This is clearly beta-level quality, it won't lose your money, but
kittens can be killed if you use this in production! Don't do that!

This payment client uses the proper P2P client and does not require
any additional indexing services and therefor is very much geared
towards the most private it can be with SPV, while staying scalable and
nible.

The only GUI available right now is for the desktop, it should be pretty
simple to re-do a nice GUI for some mobile or similar form-factors.
That is one area that QML simply rocks in.