Be smarter about blockheight
This solves the issue where we expect too many blocks when a client is
restarted while being at the tip and there is more than 300 seconds
passed since the last block.

This now doesn't show we are waiting for blocks until an actual block is
mined by having a 25min variance allowance.