Move total balance for better UX
User testing showed that the current-wallet-balance followed with the
total-balance was confusing, at best people mixed them up.

This moves the total balance to be on the top-right as to be away from
the balance of a single wallet to avoid this confusion.
Should there be space problems we show a field in the left flickable which
is now always the top one, making sure that the wallet specific list of
properties is no longer intersected by the total-balance one.