Split HD key managent into created and bloom-asked
The bloom filter didn't include all created keys and thus we didn't send
the bloom in some cases.

Now we keep track of exactly which keys were sent and we re-send a bloom
filter if half of those have been used in order to not miss out on any
transactions.