Start payment details model
To allow the UI to modify the payment we split it into details. A detail
is, for instance, an output. This allows the user to add multiple
outputs in one payment, and thus one transaction.

Further payment details can be an input selection (to select exactly
which coins need to be spent) and others may be added later.