Move most of the payment complexity into cpp
This moves all the checks and UI interactions into the model, making the
QML much lighter and less error-probe.

At the same time this allows us to add new UX features like when a 'max'
button is pressed it live-updates the amounts in the fiat and bch fields
as the user changes any spending in other outputs.