Turn off data for older Qt
The problem of all outputs and inputs having zero space is now simply
worked around by not generating that data for older runtimes of Qt.
This makes it look Ok, but we indeed do not show the inputs/outputs
in the UI unless the user is running Qt 5.15 or newer.