Avoid memory-management issues
FloweePay is an appliction singleton and as such a global that is
deleted outside of main().

The FloweePay app owns most of the data, including (implicitly) the
private keys. Private keys are managed with a LockedPoolManager in
order to enable page-locking and avoid them being swapped out.
LockedPoolManager is also a singleton, and so we make sure that it
is created _before_ FloweePay and as a result deleted after FP
releases the memory managed by the LPM.

Singleton interactions are always a bit messy.