Fix the showing of the transactions.
This finally fixes the odd case where a default wallet that is fully
encrypted did not show any transactions directly after the user typed
the password.
Turns out this is some weird cornercase in Qt/QML and removing some
code fixes this.
The downside; we now show a qml warning on startup. :shrug: