[Android] save on hiding the app.
It turned out to be impossible to save more than a couple of KB on exit,
so we moved that to the app becoming inactive.

On Android, unlike desktop, exiting the app is a 2-stage user interaction.
First the app is made inactive (equivalent to the minimize on desktop) and
at any point in time later the app is closed or killed.

So, while we save on exit on desktop this doesn't work on mobile and we must
save on it becoming inactive in order to make sure we don't lose data.