Use member instead of global
The chain is set before we init using a global (static, really).
What we really use is set at init as a member, while the global
can change later but is ignored.
So we should use the member to avoid any weird code updating the global
after the init point, to something we are not actually using.