Use namespaces
This works around an inconsistency in the QML language.
Inside a file variables are preferred local, hiding the global
ones that may have the same name.
The same is not true for named QML files (classes). An imported
is used before one of the same name in the same directory.

To avoid future additions to Qt Controls breaking our code, and
to generally avoid surprises with common names, lets standardize
using namespacing for Qt-Quick-Controls classes.