Replace blockheight (number) with date selector
Importing a key as private as possible means we don't check some central
indexing server for the first time an imported key or seed is used.
But, to avoid scanning the entire blockchain, we go back only as far
as the first usage that the user remembers.
Now, instead of asking a blockheight (which most users have no clue
about) we now ask for a year / month combo and start scanning from
the block at the beginning of that timespan.