Avoid save-file writing
On saving, check if the created data is the same as the data on disk,
and avoid writing if so. Especially on a phone overwriting the same file
every time is a bad idea. So this avoids that.

Additionally, we now move saving from shutdown to the when the user
makes the app inactive. This should make actual shutdown not need to
save in almost all cases.