Refactor WalletHistoryModel
Instead of storing the internet data in reverse, we now simply take the
view-requests and revert those.
This makes the amount of places where we have to account for the
reversal much less (now just one) which helps maintainability.

Additionally, bring back the section-is-an-enum idea to avoid a
string-malloc for each row when one per section is enough.