Finetune how we respond to bad QML.
As final deployments never had bad QML, this should be seen as a tool for
developers and for debugging.
So, according to the fail-fast concept we now abort() on detection of
unparsable QML.
Additionally we log the line, which means we can see the result on
mobile should it have logging compiled in.