Separate app singleton creation and p2p init
As we need to prepare the blockheaders file before we load the p2p net,
it is much cleaner to separate that from the construction of the
singleton.
This also allows us to move the creation of the network logger to an
earlier point in time to also be able to log the startup sequence.