Improve account selector in menu
Working with the UX guy we massively improved the account selector.

Instead of trying to shoe-horn the 'add account' under a non-existing
account we realized it fits better on its own. Even the moving of it
after adding one will not cause too big a problem due to the
different look afterwards.

The actual opened menu will also behave better if you have more than
one account in this new design and more predictable.