Re-use secrets view in mobile 'backup' screen.
This moves the desktop account-details specific WalletSecretsView
component to live in the common area and we then use it from the backup
view for a wallet in mobile.

The only change is that we automatically detect if the content is too
wide to fit and we split it over 3 lines instead of 2.