Add delay on camera init
Move to the next eventloop event the usage of the camera (finding out
resolutions etc) which looks to help stability.

Also remove the init-on-startup again, which had the nasty side-effect of
turning on the camera for a very brief time due to some Qt bug that
ignores the 'active' boolean on the camera object.