Fix initial save of private wallet data
This makes sure that after initial setup of the wallet we save the
private data and the wallet that was not owned by the user is stored as
such.

On Android the destructor often is not properly called, which leads to
us not saving this on shutdown and that showed this bug.