[Mobile] Make 'to top' button appear earlier.
The button to move the view to the top of the list is hidden when we are
at the top of the list. This new version makes the appearing if the
button happen based on the amount of rows we scrolled down, making it
appear faster and be more predicatable.