Iterate on account history look.
This show the bch price on the front page too and shows 'moved' as a
label in the popup.
We give the full width to cashfusion transactions, no point is listing
the amount of sats fee paid.

This always puts a plus or minus in front of the price indicating the
receiving vs sending of funds.