Make model a bit more robust in an async world.
Since recently the wallet can delete stuff, we need the model to
be more robust in accepting data not being there.

The "async" statement refers to the fact that the wallet acts
one messages from the network, which may be in any thread.
The signals from the wallet get handled by the WalletHistoryModel in the
Qt-UI-thread, asynchronous.