UX improvements to prioritize normal usage.
When looking at the real usage of flowee pay, we aim for this
to be used 90% (or more) of the time as a better wallet. Wallet
in the physical sense.

This implies that the vast majority of usage is going to be about
paying or receiving.
A small number of times you'll use it to look up the history of
payments.

As such, its Ok to downgrade the account history and make the main
usage more prominent and easy to find.