Make the button not the way to stop editing.
When the user starts editing the text label, the way out is to use the
(virtual) keyboards 'enter' or 'done' button.

We mark this button as disabled while editing to avoid weird situation.

This also works around the application completely hanging in Qt651 on
the phone. Hanging as in: Android suggests it to be killed.