Avoid sending duplicate filters.
The backend will tell us that there is good reason to believe the remote
needs to get a new bloom filter, which is causing a lot of extra traffic
if there actually is no change in the utxo of the wallet.

This introduces a bool to mark changes that would affect the bloom
filter to be send, and only upload a new one when that bool is true.