Work on basic blockheaders changes
The flowee lib added support for various blockheader improvements.
An important one is the ability to start from a checkpoint and thus
lower the requirements at the cost of a slightly lower security.
This adds support for that.

The local part is that we stop shipping the 'info' file (some 25MB
for the full chain) in the APK / deb files, instead buiding it on first
start.