Re-work 'back' button behavior
Pressing 'Escape' or (android) back button is now more logical.
Closing the menu is new, going back to the 'main' tab as well.
This also fixes some odd behavior when using the back button while the
camera is active.

And last, when there is nothing more to back out of, on Android we end
up closing the application. This is what people expect on that platform.