[Android] when ipv6 available, connect to such peers
This adds (the first!) an actual Java class to do the checking which
interfaces are available and we then instruct the AddressDB to pick
addresses matching that.

In other words, when the Android device has a functional IPv4 network
interface, we will try to connect to peers on that IP version.
Same with IPv6.

Both can be active at the same time.