Work harder to make headerSyncComplete get called
A wallet might get stuck in silly situations like starting up when
there is no block for a long time and then not actually being online
at the moment a new block gets mined.

This fix makes sure that at startup we 'unstuck' such a wallet without
there being a need for a new block to be mined while the app is running.