Add asserts and avoid duplicate work
When a wallet is updating we may get a lot of new transactions added in
a short amount of time, this avoids calling updateMap for each one in
turn and instead triggers an update only if there isn't one in-flight.

This takes advantage of the fact that the wallet does its update in a
different thread than the UI-one.