Small UX fixes
The main font on Android seems to not ship with 'semi-bold' on
various devices. Which means that the title just was the same
as the content and while I like subtlety, in this case if failed.
Better to go with bold than with no difference at all...

Additionally small spacing / sizing fixes.